third_party/WebKit/LayoutTests/inspector/sass/test-ast-scss-7.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="./sass-test.js"></script>
<style>
pre {
    font-family: monospace;
}
</style>
<script>

function test()
{
    InspectorTest.evaluateInPage("getSASS()", onSASS);

    function onSASS(result)
    {
        InspectorTest.parseSCSS("", result.value)
            .then(InspectorTest.dumpAST)
            .then(InspectorTest.validateASTRanges)
            .catch(console.error.bind(console))
            .then(InspectorTest.completeTest);
    }
}

function getSASS()
{
    return document.querySelector(".snippet").textContent;
}

</script>
</head>

<body onload="runTest()">
<p>
Verifies AST of SCSS with pseudo classes
</p>
<pre class="snippet">
/**
 * Remove inner padding and border in Firefox 4+.
 */
button::-moz-focus-inner,
input::-moz-focus-inner {
  border: 0;
  padding: 0;
}
</pre>
</body>
</html>
