third_party/WebKit/LayoutTests/inspector/sass/test-ast-scss-7-expected.txt

Verifies AST of SCSS with pseudo classes

/**
 * Remove inner padding and border in Firefox 4+.
 */
button::-moz-focus-inner,
input::-moz-focus-inner {
  border: 0;
  padding: 0;
}
=== AST === 
rule 0: "button::-moz-focus-inner,
input::-moz-focus-inner"
    property 0
        name: "border"
            range: {"startLine":5,"startColumn":2,"endLine":5,"endColumn":8}
        value: " 0"
            range: {"startLine":5,"startColumn":9,"endLine":5,"endColumn":11}
        range: {"startLine":5,"startColumn":2,"endLine":5,"endColumn":12}
        disabled: false
    property 1
        name: "padding"
            range: {"startLine":6,"startColumn":2,"endLine":6,"endColumn":9}
        value: " 0"
            range: {"startLine":6,"startColumn":10,"endLine":6,"endColumn":12}
        range: {"startLine":6,"startColumn":2,"endLine":6,"endColumn":13}
        disabled: false
======
Ranges OK.
